pyproject.toml
[project]
name = "myapp"
version = "0.1.0"
description = "Knowledge Graph CV Builder using Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30.0",
    "google-generativeai>=0.3.0",
    "streamlit-agraph>=0.0.45",
    "python-dotenv>=1.0.0",
]
